requirements.txt
Flask==0.10.1
gunicorn==19.3.0
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.10.4
python-dateutil==1.5
twython==3.2.0